requirements.txt
py-cord==2.4.1
pydub==0.25.1
python-dotenv==1.0.0
py-cord[voice]==2.4.1